backend/requirements.txt
Flask
flask-cors
python-dotenv
pydantic
python-multipart

asyncio
mcp
google-genai